Data.BitSet.Generic

Description

A space-efficient implementation of set data structure for enumerated data types.

Note: Read below the synopsis for important notes on the use of this module.

This module is intended to be imported qualified, to avoid name clashes with Prelude functions, e.g.

 import Data.BitSet.Generic (BitSet)
 import qualified Data.BitSet.Generic as BS

The implementation is abstract with respect to container type, so any numeric type with Bits instance can be used as a container. However, independent of container choice, the maximum number of elements in a bit set is bounded by maxBound :: Int.

Bit set type

data GBitSet c a

A bit set with unspecified container type.

Operators

(\\) :: GBitSet c a -> GBitSet c a -> GBitSet c a

O(max(m, n)). See difference.

Construction

empty :: (Enum a, Bits c, Num c) => GBitSet c a

The empty bit set.

singleton :: (Enum a, Bits c, Num c) => a -> GBitSet c a

O(1). Create a singleton set.

insert :: a -> GBitSet c a -> GBitSet c a

O(d). Insert an item into the bit set.

delete :: a -> GBitSet c a -> GBitSet c a

O(d). Delete an item from the bit set.

Query

null :: GBitSet c a -> Bool

O(1). Is the bit set empty?

size :: GBitSet c a -> Int

O(1). The number of elements in the bit set.

member :: (Enum a, Bits c) => a -> GBitSet c a -> Bool

O(d). Ask whether the item is in the bit set.

notMember :: (Enum a, Bits c) => a -> GBitSet c a -> Bool

O(d). Ask whether the item is in the bit set.

isSubsetOf :: GBitSet c a -> GBitSet c a -> Bool

O(max(n, m)). Is this a subset? (s1 isSubsetOf s2) tells whether s1 is a subset of s2.

isProperSubsetOf :: Eq c => GBitSet c a -> GBitSet c a -> Bool

O(max(n, m). Is this a proper subset? (ie. a subset but not equal).

Combine

union :: GBitSet c a -> GBitSet c a -> GBitSet c a

O(max(m, n)). The union of two bit sets.

difference :: GBitSet c a -> GBitSet c a -> GBitSet c a

O(max(m, n)). Difference of two bit sets.

intersection :: GBitSet c a -> GBitSet c a -> GBitSet c a

O(max(m, n)). The intersection of two bit sets.

Transformations

map :: (Enum a, Enum b, Bits c, Num c) => (a -> b) -> GBitSet c a -> GBitSet c b

O(d * n) Transform this bit set by applying a function to every value. Resulting bit set may be smaller then the original.

Folds

foldl' :: (b -> a -> b) -> b -> GBitSet c a -> b

O(d * n) Reduce this bit set by applying a binary function to all elements, using the given starting value. Each application of the operator is evaluated before before using the result in the next application. This function is strict in the starting value.

foldr :: (a -> b -> b) -> b -> GBitSet c a -> b

O(d * n) Reduce this bit set by applying a binary function to all elements, using the given starting value.

Filter

filter :: (Enum a, Bits c, Num c) => (a -> Bool) -> GBitSet c a -> GBitSet c a

O(d * n) Filter this bit set by retaining only elements satisfying predicate.

Lists

toList :: Num c => GBitSet c a -> [a]

O(d * n). Convert this bit set set to a list of elements.

fromList :: (Enum a, Bits c, Num c) => [a] -> GBitSet c a

O(d * n). Make a bit set from a list of elements.

Internal

toBits :: GBitSet c a -> c

O(1). Internal function, which extracts the underlying container from the bit set.

unsafeFromBits :: (Enum a, Bits c, Num c) => c -> GBitSet c a

O(1). Internal function, which constructs a bit set, using a given container value. Highly unsafe, because we don't check if bits in the given value correspond to valid instances of type a.